make PopupControls component based on Popup component
