+++ src/ui/components/index.js
@@ -2,3 +2,4 @@ export * from './Button';
 export * from './LevelButton';
 export * from './PageHeader';
 export * from './Popup';
+export * from './PopupControls';
